requirement.txt
about-time==4.2.1
absl-py==2.1.0
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alive-progress==3.1.5
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
autograd==1.7.0
av==13.1.0
babel==2.16.0
bark==0.1.5
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
boto3==1.35.48
botocore==1.35.48
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cma==3.2.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
coqpit==0.0.17
coqui-tts==0.24.2
coqui-tts-trainer==0.1.4
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dataclasses-json==0.6.7
datasets==3.0.1
dateparser==1.1.8
debugpy==1.8.7
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.30.3
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.2
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
funcy==2.0
gradio==4.44.1
gradio_client==1.3.0
grapheme==0.6.0
greenlet==3.1.1
grpcio==1.67.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.23.5
humanfriendly==10.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
intel-openmp==2021.4.0
ipydatawidgets==4.3.5
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonlines==1.2.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jstyleson==0.0.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2
llama-cloud==0.1.2
llama-index==0.10.68
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-huggingface==0.2.3
llama-index-embeddings-openai==0.1.11
llama-index-embeddings-openvino==0.2.1
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post3
llama-index-llms-huggingface==0.2.8
llama-index-llms-openai==0.1.31
llama-index-llms-openvino==0.2.2
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==2.2.0
mistune==3.0.2
mkl==2021.4.0
modelscope==1.19.0
modelscope_studio==0.5.0
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
nncf==2.13.0
notebook_shim==0.2.4
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
onnx==1.16.2
openai==1.51.2
opencv-python==4.7.0.72
openvino==2024.5.0.dev20241024
openvino-genai==2024.5.0.0.dev20241024
openvino-telemetry==2024.1.0
openvino-tokenizers==2024.5.0.0.dev20241024
optimum==1.23.1
optimum-intel==1.20.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==2.0.0
pydub==0.25.1
Pygments==2.18.0
pymoo==0.6.1.3
pyparsing==3.2.0
pypdf==4.3.1
pypinyin==0.53.0
pyreadline3==3.5.4
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.12
pythreejs==2.4.2
pytz==2024.2
pywin32==308
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
qwen-vl-utils==0.0.8
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.20.0
ruff==0.6.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.2.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
stack-data==0.6.3
starlette==0.40.0
striprtf==0.0.26
SudachiDict-core==20241021
SudachiPy==0.6.8
sympy==1.13.1
tabulate==0.9.0
tbb==2021.13.1
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado==0.18.1
text-generation==0.7.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
timm==1.0.11
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.2
tomlkit==0.12.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1+cpu
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traittypes==0.2.1
transformers==4.42.4
typeguard==4.3.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.31.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
